Revert "Check isReadableMemory in asserts"
This reverts commit a601abda05a0a7096806c9812fb85d7da71eebfa.
