+++ pkgs/objective_c/ffigen_c.yaml
@@ -20,7 +20,6 @@ functions:
     - 'disposeObjCBlockWithClosure'
     - 'isValidBlock'
     - 'isValidObject'
-    - 'isReadableMemory'
     - 'Dart_InitializeApiDL'
   leaf:
     include:
