+++ pkgs/objective_c/lib/src/c_bindings_generated.dart
@@ -126,11 +126,6 @@ external bool isValidBlock(
   ffi.Pointer<ObjCBlock> block,
 );
 
-@ffi.Native<ffi.Bool Function(ffi.Pointer<ffi.Void>)>(isLeaf: true)
-external bool isReadableMemory(
-  ffi.Pointer<ffi.Void> ptr,
-);
-
 typedef ObjCSelector = _ObjCSelector;
 
 final class _ObjCSelector extends ffi.Opaque {}
